making KWrite style guide compilant...
I've removed the broken KGuiCommand
stuff and ported the frontend to KAction/XML-GUI.
KWrite is now translatable for non-european
languagues (char* everywhere for translated
strings), has the standard toolbar and menu
layout and shares the keyboard shortcuts
with the rest of the desktop.
Not exciting, but essential for a uniform
look&feel of KDE.

The backend part has a broken design now, but
it was very messy before, so this is ok I
think. Someone has to rewrite this (I *don't*
volunteer :-)

Fixes the following bugs:
4272, 4273, 5046, 4281, 4819, 2490, 4811, 4901
and 5041.

TODO:
porting the bookmark stuff (not compilated)
merge the cluttered configuration dialogs,
make them translatable. (char* here too. --
*arghh*)
Fixing editing of Unicode-documents (#868,#2478).
I hope this is just a bug :-/

I will fix the bookmarks and the dialogs and
inform the original author.

Christian Gebauer