GUI:
use ctrl-d, then you can type stuff such as:
s/needle/replacement/ -- to run this sed-like command on the current line
%s/needle/replacement/ -- to do it on the whole file
%s/needle/replacement/n -- to do it on the whole file, case insensitively
$s/needle/replacement/ -- to do it to the selection

While the core is in, and it does actually do the replacement internally,
the code is so undocumented that it never actually sets the text in the
document :)  Cullmann has "volunteered" to fix this.  Rather, this commit log
has volunteered him.

Also supported is char:number, where the number can be
0-255 (dec), 00-0377 (octal), 0x0-0xFF or x0-xFF for hex.  It can also be
a 2 byte number, in the same format, in which case it'l treat it as unicode.
Yes, i18n people, I'm an equal opportunity coder :)