Summary: Misc. bugfixes, approved by Christoph
Selection re-painting improvements, now only repaints as necessary
Squashed the selection starting from the wrong place bug (hooray!)
Fix endCol value - same problem as when I fixed the endX value in the last commit
Re-worked the scrollbar appearing and disappearing code for non-word-wrap.  This is the "proper" way to do it.  The only thing to do for this now is to change a few -- and ++ to -/+ scrollbarHeight / fontHeight (how do we get the scrollbar height in a non-deprecated way? Can't find it in the docs)
Fixed cursorCoordinates(), now codeCompletion is displayed properly in word-wrapped mode (and normal too, we weren't taking into account the leftBorder->width().
Highlight only the viewLine with the cursor, not the complete real line
Create a tagLines() function that takes cursors for efficiency in word-wrapped mode
Fix mouse selection bug where x could be negative in placeCursor()
Fix resizeEvent to updateView when the height has increased in word-wrapped mode, and to scroll back if startX > 0 in non-word-wrapped mode
Clean up some repeated code (new LineMaxX & LineMaxCol functions)
Move LineRange to KateDocument so that it can be passed to paintTextLine, in preparation for optimisations to paintTextLine