Another big commit, this one was unavoidable (much code shuffling)
As foreshadowed on kwrite-devel, this is the separating of the end stages
of rendering into KateRenderer, and the genericising of most of the decoration
details into a single KateAttribute class. (Maybe one day Kate might not be
a bunch of monolithic classes... here's hoping).

A possible benefit of this includes bidi support, I'm evaluating how difficult
this would be (hopefully not that much!).

I've also merged in a patch (with some changes) from Anakim Border
<aborder@sources.sourceforge.net> which makes dynamic word wrapping respect
the initial indentiation of the line in drawing further lines.  This is not
yet customisable (that is on my todo list).  I should have done this seperately,
but it's done now.

This brings much more flexibility to the rendering; already support is there
for additional decorations to be used in syntax highlighting, except that the
options need to be added to the definition loading.  More rendering features
are planned.

I have hunted down all the bugs I can find, so it should be safe to use, but
the redraw performance is temporarily down (mostly not noticeable unless you are
doing heavy navigating).  I have a few hunches as to where
it went (I removed the KateFontMetrics class and its caching, hoping QFontMetrics
was up to the abuse) and it will be hunted down soon.

I also have more ideas on further performance improvements not related to this
patch, stay tuned :)

Feedback welcome...

CCMAIL:aborder@sources.sourceforge.net
CCMAIL:56214-done@bugs.kde.org
CCMAIL:kwrite-devel@mail.kde.org