External Tools feature ====================== This provides a menu, Tools->External Tools that can contain user defined commands for launching external applications. The commands can contain macros that is expanded to use data fom kate in the arguments: %URL - URL of the active document %URLs - a list of URLS of all doen documents %directory - the directory of the current document %filename - the filename of the current document %line - the line of the cursor in the current document %column - the column of the cursor in the current document %selection - the selected text of the current document %text - the text of the current document
A page is added to the config dialog to manage the menu and commands.

A few sample commands are provided.

The actions for the commands are displayed in the shortcuts dialog, and
shortcuts are saved in the externaltools data file.

TODO: make use of document mimetypes to limit usage, some more checking
of the commands, some adjustments of dialogs, cleaning up of data file.