fix Bug 78652: word wrap breaks long TeX commands:  o added two methods:         o bool KateHighlighting::canBreakAt(QChar c, int attrib) to determine 	  if a line can be broken at this character (which has attribute attrib).         o QString readWordWrapConfig() reads the wordwrap deliminators.
 o use m_additionalData so that the wordwrapping works with imported
(included) rules as well (Anders:It really works now, my test.html file
contained a subtle mistake).

How to use this?
In the keywords tag of the XML file (in the general section) add a
'wordWrapDeliminator' attribute. For example (latex.xml):
<keywords weakDeliminator="\" wordWrapDeliminator=",{}[]"/>
lines now break at the ,{}[] characters and at a space.

If no wordWrapDeliminator attribute is specified it defaults to:
\t.():!+,-<=>%&*/;?[]^{|}~\\ (the standard deliminators).