More (CStyle) indent fixes from Jesse:
Current fixes:
- fixes some minor doxygen issues
- fixes runaway cursor when aligning an empty line
- fixes nesting of if/for/while in the forward-sloping direction
- prevent the letter 'n' from triggering cursor movement unless absolutely necessary
- align ending ) with opening ( if its the only thing on the line
- added one option to indentation config page to configure doxygen auto-insert of leading "*" (as
well as a variable called 'auto-insert-doxygen')
- makes sure the Align menu action obeys ReadOnly mode.

For the Doxygen option, I hijacked another KateDocumentConfig option. This is not where it belongs.
But it was the simplist place and required the fewest amount of changes. How the indenters will
properly' affect config is still to be determined. The KateDocumentConfig options will be 'cleaned'
for 4.0 anyhow.

Static wordwrap behavior where a wordwrapped line is then indented is not in this patch and won't
be
in for 3.3.  It worked quite well in a few normal scenarios but still had problems. It's too late
in the game for me to risk it. Sorry :(