Dont use hardcoded size for the filelist icon, as the size of KIconSmall may configurable. Add read/write config in file list, and won group to the config file.
And finally, the shaded backgrounds.
The filest will shade the background of viewed/edited files with a color, so that
the most recent files have a stronger shade. edited files gets a second color blended in.
The colors can be configured in the new 'Document List' panel in the config dialog, and
the feature can be turned off (but please try it on for a while and let me know about your
thoughts, how well does this work? Can you come up with nice default colors?)