The feature years in the promising has finally landed... arbitrary syntax highlighting is available from the ktexteditor interface + now hooked up all the way to rendering.
Needs testing, however.

Also run into a problem with QTextLayout: selection decorations which
i've been abusing to do interesting stuff with doesn't seem to use some
decorations such as bold (i guess this is so the layout is not
affected).  This will mean more thinking will need to be done as to
whether relayouting has to occur for dynamic stuff, etc, or whether bold
will just not be available to dynamic highlighting while Qt doesn't
support it.