Tab width is independent from indentation width.
With this modification the tab width no longer dictates indentation when
"Insert spaces instead of tabulators" is not used; instead, indentation is
always by the "Number of spaces", but a mix of tabs and spaces if the tab width
does not equal the indentation. Consequently, if one wants only tabs for
indentation, it is necessary to set the indentation equal to the tab width.

For this reason, the option "Use spaces to indent" is removed.

Users who use "Insert spaces instead of tabulators" will not notice any change.