avoid flicker when changing the active document.
the qt4 documentation claims that using 
mainWindow()->toolBar()->setUpdatesEnabled (false); is enough, but in 
practice, we need to disable painting on the window, hide the toolbar, 
do the "drawing" and then show the toolbar again, and enable painting on 
the main window.

i also fixed some indentation problems with this function.