Fix a kate crash when closing it and having loaded the kate's symbols viewer plugin.
The problem was that there was a nasty casting from KatePluginSymbolViewerView to Kate::PluginView, but KatePluginSymbolViewerView didn't inherit Kate::PluginView; so that when Kate::PluginView::writeSessionConfig() was called, the method was not found and kate crashed.

I tried first to just make KatePluginSymbolViewerView inherit from Kate::PluginView, but that doesn't work because that would be inheriting from two QObject subclasses (it inherits from KXMLGUIClient already) and that's not allowed by Qt.

So the solution was just creating a KatePluginSymbolViewerView2 and having KatePluginSymbolViewerView as a delegate. Can I backport this fix to KDE 4.0?

CCMAIL:kwrite-devel@kde.org