I've noticed that in KDE4, the "Find in Files" kate plugin doesn't have directory completion anymore for the "Folder:" combo box.
In the code I see cmbDir->completionObject()->setMode(KUrlCompletion::DirCompletion);. So in theory, it does have directory completion. I looked at the code in the Filesystem Browser to see how does it get completion working, and it seems that they manually create the completion object. I've done that, and it works.

Looking at the code of KCompletion* KCompletionBase::completionObject(), it's quite clear to me that what was happening is that completionObject() created a KCompletion object intead of a KUrlCompletion object.

CCMAIL:kwrite-devel@kde.org