Literate haskell syntax highlighting bug fixes
Patch by Bjorn Buckwalter:

The change on line 330 prevents the context stack from growing
indefinitely with interleaved "literate-normal" and "normal" contexts.

The change on line 346 pops twice at the end of an
"comment_single_line" in order to return to the "literate-normal"
context (rather than the "normal" context). Without this fix the next
line starts in "normal" context which causes the bird-track to not to
be "special" in code like:
> -- My single line comment
> f = undefined