don't give focus to wrong view
don't use an anonymous single shot timer for managing the closing of the
command line editor. KateCmdLineEdit::hideLineEdit() will give focus to
it's m_view, which after changing views will be the wrong one.
instead, keep a qtimer pointer and connect the view's focusout signal to
the timer's stop slot.

BUG: 182398