Add support for character encodings in spellchecking.
For example, in LaTeX special characters like 'ä', 'ü', 'ö' can also be
represented with the commands '\"a', '\"u', '\"o'. Support for recognizing
these commands is now added (for LaTeX only so far).

The encodings are provided in the highlighting files as follows:

<spellchecking>
    <configuration encodingReplacementPolicy="EncodeWhenPresent"/>
    <encodings>
      <encoding character="&#228;" sequence="\&#34;a" />
      <encoding character="&#246;" sequence="\&#34;o" />
      <encoding character="&#252;" sequence="\&#34;u" />
      <encoding sequence="''" ignored="true" />
    </encodings>
  </spellchecking>

The attribute 'encodingReplacementPolicy' specifies what should happen with
characters for which an encoding is provided in words that replace misspelled
words. Possible options are so far: 'EncodeAlways', 'EncodeWhenPresent', or
'EncodeNever'. 'EncodeWhenPresent' only encodes the characters when the
misspelled word contains an encoded character.

The 'ignored' attribute allows to specify that a sequence of characters should
be ignored for spellchecking.

CCMAIL: kwrite-devel@kde.org
CCMAIL: kile-devel@lists.sourceforge.net