improve unit testing framework by cutting compilation time by half
Unit tests now can only be run if the flag -DKDE4_BUILD_TESTS=ON is specified in the cmake command line.
This will export the symbols marked with KATE_TESTS_EXPORT in katepart, such that the unit tests can link to the katepart directly.
If the flag is omitted, these symbols will NOT be exported, which is the old and usual behavior.

testkateregression has been moved into the tests directory and is only build if unit tests are enabled.
Linking to the katepart directly, it eliminates the need to compile the katepart twice.

CCMAIL: kwrite-devel@kde.org