SVN commit 1112314 by mwolff:
Also add closing pair for [ and ( automatically. But when we type the closing
pair manually, eat it so we don't end up with duplicated parens.

KISS principle at it's best, imo it works like a charm.

Thanks to André Stein for the patch!

CCMAIL: André Stein <andre.stein@rwth-aachen.de>