document removal before signal emission
Now the document is removed from the document model before the
documentDeleted signals are emitted. The signals trigger view updates,
which access the item list, which again accesses the already deleted
document and it's members if the order is the other way round.

Added warnings that the pointer must not be dereferences in slots
connected to the signals, because the document has already been deleted,
the pointer may only be used for lookup in hashes or maps, as long as
the hash function does not access the invalid data. The save way is to
act on the willBeDeleted signals.
This hopefully fixes BUG: 230426