initial config file, attribute expansion
In the config file attributes per tag can be specified which should
always be added during expansion, even if not explicitly specified.
Attributes can now be spacified in css selector like ways. The dollar
operator is not yet supported. There are no spaces allowed between
attribute name the qual sign and the attributte value. If the value is
not quoted it has to be an identifier or a number. The parser is about
as picky as the one of the zencoding stuff, from what experiments with
both show