emit selectionChanged() on selection mode change
Emit selectionChanged() according to the KTextEditor::View api
documentation also if there is no selection around. This is needed,
as e.g. the Kate App status bar uses this signal to update the state
of the selection mode (block selection, line based selection)

Imo, a nicer way would be an additional signal
  signal KTextEditor::View::selectionModeChanged(KTextEditor::View *view)
Maybe we should add this signal?