SVN commit 1148405 by zwabel:
Revert r1147043 (change default behavior of matchItem).
This API exists for multiple minor revisions, and you cannot simply change the default-behavior and just update the comment, as that is not compatible, and will may cause random behavior changes when transitioning between 4.4 and 4.5.
Please don't make the problem worse than it already is!
CCMAIL: mail@milianw.de