SVN commit 1152789 by zwabel:
Add a cache for MovingRanges that are contained only by a single line, and use it to speed up the rendering of highlighting-intensive documents (like in KDevelop4) by magnitues.
Please review and test this intensively for backporting, it seems to work fine.
CCMAIL: kwrite-devel@kde.org
CCMAIL: kdevelop-devel@kde.org