SVN commit 1190641 by mentasti:
fixed a bug that clear all displayed sql results when removing connection, even if the connection is not related to the current resultset. Now, to avoid query resource leaks, the displayed resultset is emptied if the associated db connection is removed, otherwise, data persists..
added "clear resultset" button. this is particularly useful if you want to release a table.
fixed a tableview rendering bug when cleaning resultsets