KTextEditor Plugin: highlight selection
It highlights all occurances of the current selected word. It only
highlights whole words. It kind of is similar to CTRL+H, but CTRL+H
jumps to the next occurance. In this case, the moving ranges are reused.

The *ui.rc file is not installed by intention (no gui atm)

CCMAIL: kwrite-devel@kde.org