Make konsole current doc sync work in some repls
The Konsole plugin for kate is very usefull,
especially when working with a language that has a
read eval print loop(such as python-ipython, ruby-
irb, haskell-ghci,ect.). If you have it set to
sync up with the directory of the current document
it does it even when inside of a repl, or it tries
to, ipython can understand cd dir, some others
can't. ghci can understand :cd dir and you can do
Dir.chdir(dirname) for irb.